% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{define_functional_forms}
\alias{define_functional_forms}
\title{Define the functional forms for linear and neural network models}
\usage{
define_functional_forms()
}
\value{
A list containing two functions:
  \item{linear}{A linear model function that computes a linear combination of features.}
  \item{neural_network}{A neural network model function that computes a non-linear combination of features using one or more hidden layers.}
}
\description{
This function returns a list of functional forms that can be used for modeling,
including a linear model and a neural network model.
}
