% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systems.R
\name{define_systems}
\alias{define_systems}
\title{Define Systems for Simulation}
\usage{
define_systems()
}
\value{
A list of systems, where each system is a list containing:
  \item{name}{Name of the system}
  \item{description}{Description of the system's structural equations}
  \item{data_func}{A function that generates data for the system}
}
\description{
This function defines a list of systems for generating simulated data. Each system 
represents a different structural equation model (SEM) that can be used to simulate
non-linear relationships with or without external actions. The systems vary in complexity,
ranging from simple non-linear models to complex high-dimensional models.
}
\details{
The available systems are:
1. Non-linear System with External Actions
2. Modified Non-linear System with External Actions
3. Reduced Complex Non-linear System with 21 Variables
4. Complex Non-linear System with 50 Variables

Each system contains a `data_func` function that generates a dataset of a specified size.
These datasets are created using a combination of random noise and external actions
depending on the environment specified.
}
\examples{
systems <- define_systems()
# Generate data from the first system
data <- systems[[1]]$data_func(n = 1000)
}
