% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{make_predictions}
\alias{make_predictions}
\title{Prediction function}
\usage{
make_predictions(new_data, model_params, model_func, parameters = NULL)
}
\arguments{
\item{new_data}{A data frame containing the new data for which predictions are to be made.}

\item{model_params}{A numeric vector of the trained model parameters.}

\item{model_func}{A function that computes predictions based on the model parameters and data.}

\item{parameters}{A list of additional parameters for the model function (optional).}
}
\value{
A numeric vector of predictions.
}
\description{
This function generates predictions based on the provided model parameters and data.
}
