% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{simulate_data}
\alias{simulate_data}
\title{Generalized data simulation function for multiple environments}
\usage{
simulate_data(system, n, mu_A1, sigma_A1, mu_A2, sigma_A2)
}
\arguments{
\item{system}{A function representing the system used to generate data.}

\item{n}{An integer specifying the number of observations to generate for each environment.}

\item{mu_A1}{A numeric vector specifying the means of the external actions for the first variable across environments.}

\item{sigma_A1}{A numeric vector specifying the standard deviations of the external actions for the first variable across environments.}

\item{mu_A2}{A numeric vector specifying the means of the external actions for the second variable across environments.}

\item{sigma_A2}{A numeric vector specifying the standard deviations of the external actions for the second variable across environments.}
}
\value{
A list of data frames, each corresponding to one environment.
}
\description{
This function simulates data for multiple environments based on a specified system.
}
