% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{train_and_evaluate}
\alias{train_and_evaluate}
\title{Train and evaluate models}
\usage{
train_and_evaluate(lambdas, data_G1, data_G2, test_data, define_forms_func)
}
\arguments{
\item{lambdas}{A numeric vector of lambda values to use for training and evaluation.}

\item{data_G1}{A data frame containing data from the first environment.}

\item{data_G2}{A data frame containing data from the second environment.}

\item{test_data}{A data frame containing the test data for model evaluation.}

\item{define_forms_func}{A function that defines the functional forms for modeling.}
}
\value{
A data frame summarizing the results, including the MSE for each lambda and model type.
}
\description{
This function trains and evaluates models for a range of lambda values, supporting both linear and neural network models.
}
