% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{train_model}
\alias{train_model}
\title{Training function with fixed lambda}
\usage{
train_model(
  data_G1,
  data_G2,
  lambda,
  objective_func,
  functional_form,
  parameters = NULL,
  nn = FALSE
)
}
\arguments{
\item{data_G1}{A data frame containing data from the first environment.}

\item{data_G2}{A data frame containing data from the second environment.}

\item{lambda}{A numeric value representing the trade-off between MSE and CD.}

\item{objective_func}{A function that calculates the objective function.}

\item{functional_form}{A function representing the model to be trained.}

\item{parameters}{A list of additional parameters for the model function (optional).}

\item{nn}{A logical value indicating whether the model is a neural network.}
}
\value{
A numeric vector of trained model parameters.
}
\description{
This function trains a model using the provided data from two environments
by minimizing the objective function with a fixed lambda value.
}
