% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{obj_func}
\alias{obj_func}
\title{Objective function using combined MSE and CD term}
\usage{
obj_func(weights, data_G1, data_G2, lambda, model_func, parameters = NULL)
}
\arguments{
\item{weights}{A numeric vector of weights used by the model.}

\item{data_G1}{A data frame containing data from the first environment.}

\item{data_G2}{A data frame containing data from the second environment.}

\item{lambda}{A numeric value representing the trade-off between MSE and CD.}

\item{model_func}{A function that computes predictions based on the weights and data.}

\item{parameters}{A list of additional parameters for the model function (optional).}
}
\value{
The value of the combined objective function.
}
\description{
This function calculates an objective function that combines Mean Squared Error (MSE)
and Causal Discrepancy (CD) between two different environments.
}
